First working release
This is a collection of SQL Files containing either code to create stored procedures, or batches that run stored procedures. They are used to check whether data in a database is compatible with the constraints that are designed to check for bad data..
You often need this if you want to check whether a new release will synchronise with a previous version without triggering errors in constraints.